misc/HtmlTestSuite/results/test-0454.xhtml

<?xml version="1.0" encoding="UTF-8"?><html xmlns="http://www.w3.org/1999/xhtml"><head><title>&lt;link disabled&gt;'s "explicitly enabled" persists across rel changes</title>
<link href="mailto:emilio@crisal.io" rel="author" title="Emilio Cobos Ãlvarez"/>
<link href="https://mozilla.org" rel="author" title="Mozilla"/>
<link href="https://html.spec.whatwg.org/#attr-link-disabled" rel="help"/>
<script src="/resources/testharness.js"/>
<script src="/resources/testharnessreport.js"/>
<link disabled="" href="data:text/css,html { background: green }" rel="yadayada" title="alt"/>
<script>
function assert_applies(applies) {
  (applies ? assert_equals : assert_not_equals)(getComputedStyle(document.documentElement).backgroundColor, "rgb(0, 128, 0)");
}

const link = document.querySelector("link[disabled]");

async_test(function(t) {
  link.onload = t.step_func_done(function() {
    assert_applies(true);
    link.setAttribute("rel", "alternate stylesheet");
    assert_applies(true);
    assert_false(link.disabled);
  });
  link.disabled = false;
  link.setAttribute("rel", "stylesheet");
}, "HTMLLinkElement.disabled's explicitly enabled state persists regardless of rel");
</script>
</head><body/></html>
